requirements.txt
streamlit
pillow
scikit-learn
numpy